SyncOperation: split support types into separate files
Moves `SyncProgress`, `SyncPageContext<TModel>` and `DeletedPageContext`
out of SyncOperation.cs into their own files. The static `SyncOperation`
factory stays alongside `SyncOperation<TModel>`.
